+++ requirements.txt.bak
@@ -1,0 +1,31 @@
+anyio==3.6.2
+asgiref==3.6.0
+certifi==2022.12.7
+chardet==3.0.4
+charset-normalizer==3.1.0
+dnspython==2.0.0
+exceptiongroup==1.1.2
+fastapi==0.93.0
+h11==0.14.0
+httpcore==0.17.3
+httpx==0.24.1
+idna==2.10
+iniconfig==2.0.0
+Jinja2==3.0.1
+MarkupSafe==2.0.1
+mccabe==0.7.0
+mongomock==4.1.2
+packaging==23.1
+pluggy==1.2.0
+pycodestyle==2.10.0
+pydantic==1.10.5
+pymongo==3.11.4
+pytest==7.4.0
+requests==2.24.0
+sentinels==1.0.0
+sniffio==1.3.0
+sqlparse==0.4.3
+starlette==0.25.0
+termcolor==1.1.0
+urllib3==1.25.11
+uvicorn==0.20.0
