+++ src/config/Config.py
@@ -3,7 +3,7 @@ SLACK_WEBHOOK_URL = "https://hooks.slack.com/services/xxxxxxxxxxxxxxxxxx/xxxxxxx
 TELEGRAM_TOKEN = "xxxxxxxxxxxxxxxxxxxxxxxx"
 TELEGRAM_CHAT_ID = "xxxxxxxxxxxxxxxxxxxxxxxxxx"
 
-DB_HOST = "127.0.0.1"
+DB_HOST = "mongo" # mongo
 DB_PORT = 27017
 DB_NAME = "MonitoringSubdomain"
 COLLECTION_NAME = "subdomains"
