+++ src/service/functions.py
@@ -15,7 +15,7 @@ def notification_template(subdomain):
 
 def custom_logger(name):
     formatter = logging.Formatter(fmt='%(asctime)s %(levelname)-2s %(message)s', datefmt='%Y-%m-%d %H:%M:%S')
-    handler = logging.FileHandler('logs/logs.log')
+    handler = logging.FileHandler('../logs/logs.log')
     handler.setFormatter(formatter)
     screen_handler = logging.StreamHandler(stream=stdout)
     screen_handler.setFormatter(formatter)
